Cancel Ticket & Search User

+++ app/Http/Controllers/ManagerController.php
@@ -16,7 +16,7 @@ class ManagerController extends Controller
         return view('manager.home');
     }
     public function profile(){
-        $user = userinfo::where('username','Mortujaii')
+        $user = userinfo::where('username','fahim')
         ->first();
         
         return view('manager.profile')->with('user',$user);
@@ -47,6 +47,7 @@ class ManagerController extends Controller
                 'address' => $req->address
                 ]);
 
+        session()->flash('msg','Profile Updated Successfully');
         return redirect()->route('manager.profile');
     }
     public function changepass(Request $req){
@@ -74,16 +75,15 @@ class ManagerController extends Controller
             $user = userinfo::where('username',$req->uname)
                 ->update(['password' => md5($req->password)]);
 
-            $msg = "Password Changed Successfully";
+            session()->flash('msg','Password Changed Successfully');
             return redirect()->route('manager.profile');
         }
         else{
-            $msg = "Wrong Old Password";
-
             $user = userinfo::where('username',($req->uname))
             ->select('id','password')
             ->first();
 
+            session()->flash('msg','Wrong Old Password');
             return  redirect()->route('manager.changepass',['id'=>encrypt($user->id)]);
         }
         
@@ -118,7 +118,9 @@ class ManagerController extends Controller
             $tickets = PurchasedTicket::where('from_stopage_id','=',$req->fromstopage)->get();
             $users = array();
             foreach($tickets as $ticket) {
-                array_push($users,$ticket->user);
+                if(!in_array($ticket->user,$users)){
+                    array_push($users,$ticket->user);
+                }
             }
             $stopage = Stopage::all();
             return view('manager.userlist')->with('users',$users)->with('stopage',$stopage);
@@ -127,7 +129,9 @@ class ManagerController extends Controller
             $tickets = PurchasedTicket::where('to_stopage_id','=',$req->tostopage)->get();
             $users = array();
             foreach($tickets as $ticket) {
-                array_push($users,$ticket->user);
+                if(!in_array($ticket->user,$users)){
+                    array_push($users,$ticket->user);
+                }
             }
             $stopage = Stopage::all();
             return view('manager.userlist')->with('users',$users)->with('stopage',$stopage);
@@ -137,7 +141,9 @@ class ManagerController extends Controller
             ->where('to_stopage_id','=',$req->tostopage)->get();
             $users = array();
             foreach($tickets as $ticket) {
-                array_push($users,$ticket->user);
+                if(!in_array($ticket->user,$users)){
+                    array_push($users,$ticket->user);
+                }
             }
             $stopage = Stopage::all();
             return view('manager.userlist')->with('users',$users)->with('stopage',$stopage);
@@ -161,10 +167,61 @@ class ManagerController extends Controller
         //return $available_seat;
         return view('manager.flightdetails')->with('flight',$flight)->with('available_seat',$available_seat);
     }
-    public function cancelticket(Request $req){
 
-        $deleteticket = PurchasedTicket::where('id','=',decrypt($req->id))->delete();
-        return $deleteticket;
+
+
+    public function cancelticket(Request $req){
+        
+        $ticketcount = PurchasedTicket::where('purchased_by','=',decrypt($req->uid))->count();
+        
+        if($ticketcount <= 1){
+            $user = userinfo::where('id','=',decrypt($req->uid))->first();
+
+            session()->flash('msg','Ticket cannot be canceled');
+            return view('manager.userdetails')->with('user',$user);
+        }
+        else{
+            $deleteticket = PurchasedTicket::where('id','=',decrypt($req->id))->delete();
+            $user = userinfo::where('id','=',decrypt($req->uid))->first();
+
+            session()->flash('msg','Ticket cancelled Successfully');
+            return view('manager.userdetails')->with('user',$user);
+
+        }
+
+    }
+    public function searchuserlist(){
+        $users = userinfo::where('role','User')->get();
+        return view('manager.searchuserlist')->with('users',$users);
+    }
+    public function searchuserlistsubmit(Request $req){
+
+        if($req->email == "" && $req->uname == ""){
+            $users = userinfo::where('role','User')->get();
+            return view('manager.searchuserlist')->with('users',$users);
+
+        }
+        elseif($req->email != "" && $req->uname == ""){
+            $users = userinfo::where('role','User')
+            ->where('email','like','%'.$req->email.'%')
+            ->get();
+            return view('manager.searchuserlist')->with('users',$users);
+        }
+        elseif($req->email == "" && $req->uname != ""){
+            $users = userinfo::where('role','User')
+            ->where('username','like','%'.$req->uname.'%')
+            ->get();
+            return view('manager.searchuserlist')->with('users',$users);
+        }
+        elseif($req->email != "" && $req->uname != ""){
+            $users = userinfo::where('role','User')
+            ->where('username','like','%'.$req->uname.'%')
+            ->where('email','like','%'.$req->email.'%')
+            ->get();
+            return view('manager.searchuserlist')->with('users',$users);
+        }
+
+
     }
